adds startTime option to Game
this is being put in specifically for the game restoration feature in
Lightsquare, but it probably has other use cases as well.  It makes sense
to have this with the history option, as if there's already history it's
likely that the game hasn't only just started.  the two options together
should be sufficient to completely support re-starting games that were
already in progress.